move normal columns to their own file
